requirements.txt
fastapi
openai
python-dotenv
pyyaml
uvicorn[standard]
